Implemented a decorator class which intercepts NRPN and ControlChange messages from a MIDI input device, and retransmits them in at a configurable rate. Messages are dropped as needed, but the last received value for a given paramter (or control) is always guarantueed to be sent, eventually. Thus only intermediary values are dropped, if needed.
This is to improve the experience when rapidly changing values, or when changing multiple values simultaneously.

The parameter update rate setting, now affects the update rate of this decorator class, thus affecting all inputs, rather than just value changes for single parameters - which would not help when changing multiple parameters simultaneously.

This change has the added benefit of removing the timer instance in each ParameterFunction, instead replaced with a single managed thread in the decorator class.