Changed so that LastUsed setting, is only changed when a device is selected - such that if the app is opened while the last used device isn't present (and no new one is selected), we will remove the last ued device from the settings, such that the next time the app is opened and the device is present again, it will automatically be selected.
If the user selectes a different device, it will always be overwritten in settings - this only affects the case when the app is opened and no device is selected. It doesn't hurt for us to try and find the last known device, even if last time the app was opened - none was selected.